Move merged ball to a separate data folder.
